% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes-api.R
\name{.clean_geojson}
\alias{.clean_geojson}
\title{Clean GeoJSON for API Compliance (RFC 7946)}
\usage{
.clean_geojson(geography)
}
\arguments{
\item{geography}{sf object or GeoJSON list}
}
\value{
A cleaned GeoJSON list object
}
\description{
Clean GeoJSON for API Compliance (RFC 7946)
}
\keyword{internal}
