% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes-api.R
\name{.get_extract_result}
\alias{.get_extract_result}
\title{Get MobileScapes Extract Result Download Info}
\usage{
.get_extract_result(request_id, country, vintage)
}
\description{
Retrieves Azure storage details for a completed extract request.
}
\keyword{internal}
