% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes-api.R
\name{.get_extract_status}
\alias{.get_extract_status}
\title{Get MobileScapes Extract Request Status}
\usage{
.get_extract_status(request_id, country, vintage)
}
\description{
Get MobileScapes Extract Request Status
}
\keyword{internal}
