% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{.is_initialized}
\alias{.is_initialized}
\title{Check if Credentials are Initialized}
\usage{
.is_initialized()
}
\value{
Logical. TRUE if credentials are set up and valid
}
\description{
Check if Credentials are Initialized
}
\keyword{internal}
