% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes-api.R
\name{.perform_request}
\alias{.perform_request}
\title{Perform a MobileScapes Request and Parse the JSON Response}
\usage{
.perform_request(req)
}
\arguments{
\item{req}{An httr2 request object.}
}
\value{
Parsed JSON response body (as a list).
}
\description{
Perform a MobileScapes Request and Parse the JSON Response
}
\keyword{internal}
