% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes-api.R
\name{.submit_origins_extract}
\alias{.submit_origins_extract}
\title{Submit a MobileScapes Origins Extract Request}
\usage{
.submit_origins_extract(
  geofence_ids,
  start_date,
  end_date,
  days_of_week,
  time_of_day,
  dwell,
  country,
  vintage
)
}
\description{
Submit a MobileScapes Origins Extract Request
}
\keyword{internal}
