% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{get_bearer_token}
\alias{get_bearer_token}
\title{Get Valid Bearer Token}
\usage{
get_bearer_token()
}
\value{
Character. Valid bearer token string
}
\description{
Returns a valid bearer token, refreshing if necessary. This is the main
function to use when making API calls.
}
\examples{
\dontrun{
init_credentials(
  client_id = Sys.getenv("CLIENT_ID"),
  client_secret = Sys.getenv("CLIENT_SECRET")
)
token <- get_bearer_token()
}

}
