% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes_api.R
\name{get_request_results}
\alias{get_request_results}
\title{Get MobileScapes Request Results Information}
\usage{
get_request_results(bearer_token, request_id)
}
\arguments{
\item{bearer_token}{Character. OAuth bearer token.}

\item{request_id}{Character. The request ID.}
}
\value{
List with Azure storage details (storageUrl, containerName, sasToken, blobList).
}
\description{
Retrieves Azure storage information from completed request for downloading queried files.
}
\keyword{internal}
