% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes_api.R
\name{separate_large_geos}
\alias{separate_large_geos}
\title{Separate Large and Small Geographies}
\usage{
separate_large_geos()
}
\value{
To be implemented.
}
\description{
Placeholder function for separating geographies based on size threshold.
Check which geographies in a given FeatureCollection are >= 5M sq. ft
and filter them into separate data structures.
}
\keyword{internal}
