% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilescapes-api.R
\name{.download_extract_files}
\alias{.download_extract_files}
\title{Download MobileScapes Extract Files from Azure Blob Storage}
\usage{
.download_extract_files(request_id, country, vintage, output_dir = "temp")
}
\description{
Download MobileScapes Extract Files from Azure Blob Storage
}
\keyword{internal}
