% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{.quietly_get_bearer_token}
\alias{.quietly_get_bearer_token}
\title{Get Valid Bearer Token Quietly}
\usage{
.quietly_get_bearer_token()
}
\value{
Character. Valid bearer token string
}
\description{
Returns a valid bearer token, refreshing if necessary, without verbosity.
}
\examples{
\dontrun{
init_credentials(
  client_id = Sys.getenv("CLIENT_ID"),
  client_secret = Sys.getenv("CLIENT_SECRET")
)
token <- get_bearer_token()
}

}
